Produce multi-channel histograms
FLIMEvents: Add PixelPhotonRouter to route PixelPhotonEvents based on
channel.

This adds settings to enable/disable each of the 16 possible channels,
plus construction and saving of multi-channel histograms in the .sdt
file. The intensity image for display remains single-channel but is now
the sum of photons only from the enabled channels.