Rename "count" events to less ambiguous names
detection_event and time_correlated_detection_event.

We may in the future have a counter processor that outputs "count"
events. Use "detection" to name these photon-or-other-event events.

(They could be photons, but they could also be the rising/falling edges
of photons, clocks, etc., so "photon" is not a good name. Nor is "edge".
Just "event" would have been a good name if we didn't already use it for
our "events". "Physical event" is too long, and these may be simulated.)