Allow successive device events with equal abstime
Although I have never encountered this with BH SPC-150, a BH SPC-180NX
was found to produce marker events followed by a photon with the same
macrotime. I do not know how PicoQuant devices behave, but allow this
case generally (as long as we perform this validation in the decoding
processors - perhaps validation should be done separately and
optionally).