Prevent cold path from interfering with inlining
In the summarize_swabian example, these changes were sufficient to cause
the hot path (up to summerize_and_print::handle_event(detection_event))
to be completely inlined into read_binary_file::pump_events() (Apple
clang 14 arm64). This resulted in an ~1.25x (MSVC 19.36 x64) to 1.5x
(Apple clang 14 arm64) speedup on real FLIM data (840 MiB).

In general it is best to segregate cold code (error handling; handling
of rare events) into functions (marked noinline) so that they do not
clutter the hot path. This is especially true for code that performs
allocations or uses stringstream.

Here only a few processors (those used in the example program) have been
optimized.

I suspect that marking the cold functions noinline is often not
necessary (in the above MSVC test, performance was similar), but it
can't hurt so long as it is limited to cases where the function is
unambiguously cold (see comment added to 'unbatch' regarding a case
where it is not so).