Use type-erased dispatch for 'route'
The previous implementation of 'route' had a hot spot in the downstream
selection, with N poorly predictable branches for every event.

By replacing with a combination of route_homogeneous (which uses index
dispatch) and type_erased_processor, we have a single (multiway)
unpredictable branch.

(Actually there is an additional branch for out-of-range indices
(discarded events). This could be further improved by requiring routers
to return exactly -1 (or N) when discarding and dispatching to a
null_sink.)

2.4% speedup observed with flim_bruker_swabian (real data, Apple M1).