Delete split processor
It is not clear that it is needed (seeing that it is not used in the
examples so far), and if it is needed, it should be implemented using a
type-based router and the `route` processor.

Removing because we would need to add a broadcast (for
time_reached_event) if we were to keep it.