Retire ref_processor and shared_processor
In addition to ref_processor being unsafe, these have the problem of
allowing multiple upstreams to be attached at the same time. Now that we
have an ergonomic type_erased_processor and processor_context-based
side-channel access, these should not be needed.

If we need a replacement, perhaps we should consider a unique_processor
that move-constructs the downstream onto the heap.