Enforce once-only call to pump()
The alternative uniform behavior would be for second and subsequent
calls to be a no-op, but this is pointless because it would mean such
usage is always a mistake. So point out the mistake.

We may want other behaviors in the future (such as monitoring a growing
file and pumping up to the current EOF but allowing further pumping
later). Those can use a different function name (such as pump_to_eof()).

The plain pump() should guarantee flushing (in the absence of errors).