Remove abstime from bin_increment_event
No longer needed, since we have removed time recording from
histogramming. More consistent with bin_increment_batch_event (which
also no longer has a time range).

This does remove the ability to `merge` a stream of bin increments (for
example with a pixel clock), hence the change to
examples/flim_bruker_swabian. But if anything, this improved the logical
grouping in the example code.

The example now needs to buffer (in the merge) the potentially larger
time_correlated_detection_event, but it is actually likely to be the
same size (16 bytes) as the former bin_increment_event containing
abstime.