Remove noexcept from most npint functions
These functions have undefined behavior (signed overflow) and we may
add checks in the future. Behavior is defined for unsigned, but we may
add checks for that case, too.

Now that we run processors in a way that allows throwing, it is no
longer useful to mark functions with narrow contracts with noexcept.