Delete rvalue-specific from sink_events
This turned out not to be very useful, as a sink that handles an event
by rvalue ref only can be rejected by checks in upstream processors.

It is also not necessary for testing handles_const_event_v and
handles_rvalue_event_v, because those have their own tests using simpler
types.