Add overflow checks to time-correlation processors
This may turn out to be unnecessary, because time-correlated pairs are
within a known time window and user code can ensure that difftime_type
is large enough to cover the time window.

That, however, relies on non-local properties of the processing graph,
and it is better to have safe code than uncertain code.

If it turns out to be detrimental to performance, we can add a way to
(optionally) disable the check. Testing for this can be done by
disabling checks globally, for example with the Swabian FLIM example.