Use correct criteria for event list constraints
In general event handlers (handle()) will accept types that are
implicitly convertible to the accepted types. Make this be the case when
the accepted types are members of a type list.

This does not affect any of our current usage (except for the tests
updated here), nor is it likely to affect typical use cases. But it
makes behavior more uniform:

- Non-template overloads of handle(): accepts implicitly convertible.

- Template overloads of handle() constrained by
  handles_event_v<Downstream, E>: accepts types that are implicitly
  convertible to types handled by Downstream.

- Template overloads of handle() constrained by type list: accepts types
  that are implicitly convertible to types in the list (new here).

(An alternative normalization might be to only ever accept exact types,
by never using non-template overloads (using std::is_same_v for a single
event type). Will it work without too much clutter? This possibility is
left for future study.)