histogram_scans to replace histogram_elementwise
The new processor combines the functionality of histogram_elementwise
and histogram_elementwise_accumulate and is more configurable.

Also, policy settings are more independent of each other so that, among
other things, testing is easier.

The old histogram_elementwise.hpp is not removed yet because the example
programs have not been converted to use the new processor.