Release GIL in handle/flush from Python
Require the list of event C++ types upon construction of a Python
Context.

Use the requested type list to wrap the input processor, so that we have
a simpler C++ overload set (for handle()) to deal with using cppyy.

Cppyy overload sets (unlike template proxies) have __release_gil__,
which we can set to True.