Python: separate compilation from execution
Instead of creating a Context directly from a Graph, first compile a
Graph to a CompiledGraph, then instantiate an ExecutionContext from the
CompiledGraph.

Also, decouple context-based access from graph nodes, so that in the
future we can support access to auxiliary objects stored inside
processors. Thus, pass an AccessTag when specifying a tracked object.

See diff of python/examples/summarize_bh_spc.py for a quick overview of
what changed.