Simplify bin_increment_batch_journal
Slightly change interface (iteration now yields a span for every batch,
including empty ones).

Use a simpler single-array encoding. For typical (small) batches, simply
insert a size prefix in the stream of bin increments. When the batch
size does not fit in the bin_index_type (interpreted as unsigned),
insert a size_t-sized size prefix spanning multiple elements.

When bin_index_type is 8-bit, batches are small (< 255 counts), and
empty batches are rare, this does not consume any more memory than the
previous implementation.

Performance of flim_bruker_swabian on macOS arm64 did not change. This
is probably a good thing, suggesting that the journaling is not
rate-limiting.