Allow batching bin increment clusters of any size
Instead of failing with an exception when a cluster doesn't fit in the
requested bucket size, allocate as large a bucket as necessary to pass
on the large cluster.

The thinking is that it is generally easier to deal with some
performance degradation (which should be small if such large clusters
are rare, which is the expected use case) than having a hard-limit on
what can be processed. Although we expect any realistic input data to
have a limited cluster size, it is not easy to strictly guarantee that
in practice.

Note that this makes batch_bin_increment_clusters play poorly with our
recycling bucket sources: every time a large bucket is allocated, it
will stay around in circulation until the bucket source is destroyed.
This issue should be addressed separately.