Python: Rewrite to use Meson and nanobind
This does not (yet) remove all uses of cppyy.

The package libtcspc._odext ("on-demand extensions") builds C++ Python
extension modules using Meson, Ninja, and the platform compiler, and
imports them. This replaces the JIT code via cppyy.

The handling of arrays (Python buffer protocol objects) is greatly
simplified thanks to nanobind::ndarray. handle_span.cpp is removed.

A class, tcspc::parameterized_data_types, is added to eliminate the need
for code to be compiled for DataTypes.